+++ src/snowflake/cli/_plugins/spcs/services/commands.py
@@ -16,10 +16,10 @@ from __future__ import annotations
 
 import itertools
 from pathlib import Path
-from typing import Generator, Iterable, List, Optional, cast
+from typing import Dict, Generator, Iterable, List, Optional, cast
 
 import typer
-from click import ClickException
+from click import ClickException, UsageError
 from snowflake.cli._plugins.object.command_aliases import (
     add_object_command_aliases,
     scope_option,
@@ -29,9 +29,16 @@ from snowflake.cli._plugins.spcs.common import (
     validate_and_set_instances,
 )
 from snowflake.cli._plugins.spcs.services.manager import ServiceManager
+from snowflake.cli._plugins.spcs.services.service_entity_model import ServiceEntityModel
+from snowflake.cli._plugins.spcs.services.service_project_paths import (
+    ServiceProjectPaths,
+)
+from snowflake.cli.api.cli_global_context import get_cli_context
 from snowflake.cli.api.commands.flags import (
     IfNotExistsOption,
     OverrideableOption,
+    ReplaceOption,
+    entity_argument,
     identifier_argument,
     like_option,
 )
@@ -39,6 +46,7 @@ from snowflake.cli.api.commands.snow_typer import SnowTyperFactory
 from snowflake.cli.api.constants import ObjectType
 from snowflake.cli.api.exceptions import (
     IncompatibleParametersError,
+    NoProjectDefinitionError,
 )
 from snowflake.cli.api.feature_flags import FeatureFlag
 from snowflake.cli.api.identifiers import FQN
@@ -196,6 +204,51 @@ def create(
         tags=tags,
         comment=comment,
         if_not_exists=if_not_exists,
+    )
+    return SingleQueryResult(cursor)
+
+
+@app.command(requires_connection=True)
+def deploy(
+    replace: bool = ReplaceOption(help="Replace the service if it already exists."),
+    entity_id: str = entity_argument("service"),
+    **options,
+) -> CommandResult:
+    """
+    Deploys a service defined in the project definition file.
+    """
+    cli_context = get_cli_context()
+    pd = cli_context.project_definition
+
+    if pd is None:
+        raise NoProjectDefinitionError(
+            project_type="service", project_root=cli_context.project_root
+        )
+
+    services: Dict[str, ServiceEntityModel] = pd.get_entities_by_type(
+        entity_type="service"
+    )
+
+    if not services:
+        raise NoProjectDefinitionError(
+            project_type="service", project_root=cli_context.project_root
+        )
+
+    if entity_id and entity_id not in services:
+        raise UsageError(f"No '{entity_id}' entity in project definition file.")
+    elif len(services.keys()) == 1:
+        entity_id = list(services.keys())[0]
+
+    if entity_id is None:
+        raise UsageError(
+            "Multiple services found. Please provide entity id for the operation."
+        )
+
+    service_project_paths = ServiceProjectPaths(cli_context.project_root)
+    cursor = ServiceManager().deploy(
+        service=services[entity_id],
+        service_project_paths=service_project_paths,
+        replace=replace,
     )
     return SingleQueryResult(cursor)